backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
clerk-sdk==0.1.9
cryptography==45.0.4
Deprecated==1.2.18
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-json-logger==3.3.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
sentry-sdk==2.35.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.3.0
svix==1.71.0
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wheel==0.45.1
wrapt==1.17.2
